backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.14"
dependencies = [
    "email-validator>=2.3.0",
    "emails>=0.6",
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "resend>=2.19.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
